[DIST-10392] add logging functionality
1) add log enabler
