+++ lib/muffin_man/enable_logger.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module MuffinMan
+  module EnableLogger
+    LOGGING_ENABLED = true
+  end
+end
